# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "DataPorts/Data Governance",
        "description": "DataPorts/Data Governance component",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "",
    "basePath": "",
    "schemes": [
        "http", "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/prelogin": {
            "get": {
                "tags": [
                    "Login"
                ],
                "summary": "Get all blockchain participant organization",
                "description": "List all HLF organizations",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema":{
                            "$ref": "#/definitions/OrganizationstListResponse"
                        }
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
				"security": [{
					"API_KEY": []
				},{
					"AuthorizationHeader": []
				}]
            }
        },
...
        "/updatePassword": {
            "put": {
                "tags": [
                    "Login"
                ],
                "summary": "Update password",
                "description": "Update your Data Governance password",
                "parameters": [
                    {
                        "name" :"Authorization",
                        "in" : "header",
                        "type": "string",
                        "description" : "My token",
                        "required" : "true"
       
                    },
                    {
                        "name": "update password",
                        "in": "body",
                        "description": "Update password for a registered Data Governance user",
                        "schema": {
                            "$ref": "#/definitions/UpdatePasswordRequest"
                        },
                        "required" : "true"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
				"security": [{
					"API_KEY": []
				},{
					"AuthorizationHeader": []
				}]
            }
        },
        "/queryOrgsByRange": {
            "get": {
                "tags": [
                    "Users & Organizations"
                ],
                "summary": "Get all the Data Governance organizations",
                "description": "List all Data Governance participant organizations",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/getOrgsByManagerAdmin": {
            "get": {
                "tags": [
                    "Users & Organizations"
                ],
                "summary": "Get all the Data Governance organizations that you have added",
                "description": "List all Data Governance participant organizations that you have added",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/getInternalOrgByName": {
            "get": {
                "tags": [
                    "Users & Organizations"
                ],
                "summary": "Get all the Data Governance organizations based on the organization under the umbrella of the HLF organization",
                "description": "List all Data Governance participant organizations based on the organization under the umbrella of the HLF organization",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
        "/registerOrg": {
            "post": {
                "tags": [
                    "Users & Organizations"
                ],
                "summary": "Register a new organization on Data Governance",
                "description": "Data Governance:Register organization",
                "parameters": [
                    {
                        "name" :"Authorization",
                        "in" : "header",
                        "type": "string",
                        "description" : "My token",
                        "required" : "true"
       
                    },
                    {
                        "name": "Register organization",
                        "in": "body",
                        "description": "Register a new organization",
                        "schema": {
                            "$ref": "#/definitions/RegisterOrgRequest"
                        },
                        "required" : "true"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
				"security": [{
					"API_KEY": []
				},{
					"AuthorizationHeader": []
				}]
            }
        },
        "/updateOrgInfo": {
            "put": {
                "tags": [
                    "Users & Organizations"
                ],
                "summary": "Update org info on Data Governance",
                "description": "Data Governance: Update organization",
                "parameters": [
                    {
                        "name" :"Authorization",
                        "in" : "header",
                        "type": "string",
                        "description" : "My token",
                        "required" : "true"
       
                    },
                    {
                        "name": "Update organization",
                        "in": "body",
                        "description": "Update info for an existed organization",
                        "schema": {
                            "$ref": "#/definitions/UpdateOrgRequest"
                        },
                        "required" : "true"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
				"security": [{
					"API_KEY": []
				},{
					"AuthorizationHeader": []
				}]
            }
        },
        "/queryUsersByRange": {
            "get": {
                "tags": [
                    "Users & Organizations"
                ],
                "summary": "Get all the Data Governance users",
                "description": "List all Data Governance users",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "name" :"Authorization",
                    "in" : "header",
                    "type": "string",
                    "description" : "My token",
                    "required" : "true"
   
                }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "General Error"
                    }
                },
                "security": [{
                    "API_KEY": []
                },{
                    "AuthorizationHeader": []
                }]
            }
        },
...
    },
    "definitions": {
        "OrganizationstListResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "list participant organizations"
                }
            }
        },
...
        "UpdatePasswordRequest": {
            "type": "object",
            "required": [
                "previouspassword",
                "newpassword" 
            ],
            "properties": {
                "previouspassword": {
                    "type": "string",
                    "required" : "true"
                },
                "newpassword": {
                    "type": "string",
                    "required" : "true"
                }
            }
        },
...
    }
}